Safari-609.1.11.1/LayoutTests/fast/canvas/canvas-modify-emptyPath-expected.txt

This tests behaviour of path modification APIs on an empty path.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test lineTo
PASS getColor(40,40) is [0,128,0,255]
Test canvas.rect() with width = height = 0.
PASS getColor(1,1) is [0,0,0,0]
Test lineTo sequence
PASS getColor(0,0) is [255,0,0,255]
PASS getColor(50,50) is [0,128,0,255]
Test quadraticCurveTo
PASS getColor(10,10) is [255,0,0,255]
PASS getColor(50,50) is [0,128,0,255]
Test quadraticCurveTo endpoint
PASS getColor(10,10) is [255,0,0,255]
PASS getColor(99,51) is [0,128,0,255]
PASS getColor(50,50) is [0,128,0,255]
Test bezierCurveTo
PASS getColor(10,10) is [255,0,0,255]
PASS getColor(50,50) is [0,128,0,255]
Test bezierCurveTo endpoint
PASS getColor(10,10) is [255,0,0,255]
PASS getColor(99,51) is [0,128,0,255]
PASS getColor(50,50) is [0,128,0,255]
PASS successfullyParsed is true

TEST COMPLETE
